pyproject.toml
[tool.poetry]
name = "trello-sort-checklist-items"
version = "0.1.0"
description = ""
authors = ["Chris Adams <chris@productscience.co.uk>"]
readme = "README.md"
packages = [{ include = "trello_sort_checklist_items" }]

[tool.poetry.dependencies]
python = "^3.10"
trello = "^0.9.7.3"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
